Add splash Screen
To minimize waiting time when using the app, we utilize Paper-DB for the
apps state. This state has to be loaded on app startup. That can create a
waiting period for the user while the system is performing set-up tasks.
Therefore a splash screen is necessary to seemlessly transition the user
from starting to ready while performing heavy short-burst action.

- Added App.java as base application class (heavy work is done here)
- Added SplashScreen-Activity to transition to Home-Activity
- Added splash screen drawable with an outlined logo of the app
- Moved Home-Activity into Activies-Package
- Added static string to contact developer on error
- Added splash screen theme that shows splash screen drawable