Optimize Game-Cycle
Every time the Game-Cycle was done, all questions got discarded & added again.
This is a very ineffective way of doing this, as in practical terms, we were
just starting the list all over again. Also the app stability suffered if
the "Next"-Button was pressed in a fast succession. These two issues can be
tackled with one solution.

- Questions for the Game-Cycle are only added once
- On repeat the List-Iterator will be acquired once again, to start anew
- Every iteration for the game-logic runs in a sequential background thread
- We will be failing in onPostCreate if no question is available